reorganize my environment component
