+++ src/app/app-routing.module.ts
@@ -2,10 +2,11 @@ import { SavedSearchesComponent } from './features/saved-searches/components/sav
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './features/home/page/home.component';
-import { MyEnvironmentPageComponent } from './features/my-environment/pages/my-environment-page/my-environment-page.component';
 import { LoginComponent } from './features/users/components/login/login.component';
 import { RegisterComponent } from './features/users/components/register/register.component';
 import { VirtualAssistantPageComponent } from './features/virtual-assistant/page-routing/virtual-assistant-page.component';
+import { MyEnvironmentPageComponent } from './features/my-environment/components/my-environment-page/my-environment-page.component';
+
 
 const routes: Routes = [
   {path: '', component: HomeComponent},
